requirements.txt
nextcord
python-dotenv
httpx
websockets
